added democracy
